[IMP] Move logic to helpdesk_mgmt_rest_api
