+++ shopinvader_helpdesk/services/__init__.py
@@ -1,4 +1,1 @@
-from . import abstract_attachment
-from . import abstract_mail_thread
-from . import attachment
-from . import ticket
+from . import services
